refactor: routes, controllers
renamed routes-controller to controllers, introduced services, offloaded routes logic to controllers(routes dont have logic now)
